[Fix] Option 컴포넌트의 props interface 위치 변경
